fixed code coverage issue with webpack introduction
Webpack adds wrapper code to polyfill the module loading code for environments that don't support modules. This is code that we don't write, so we can't test it, yet the fact it's not covered by our tests lowers the coverage %. Specifically for files with very little code (`core.ts` & `components.ts`) it results in a very low coverage %. This commit adds a webpack loader that addresses this & updates the `karma.config.ts` to use this isntead of the default coverage reporter.

closes pr 56, issue #52