chore(): automate release process
Added script (`scripts/update-pkg-versions.sh`) to assist in the release process. Running this script compiles new versions of the library & tags the current repo with the version specified in the `package.json`. It also clones the **package-ngofficeuifabric** repo & updates it accordingly with the current library version, dependencies & changelog, then tags & pushes it to it's remote. See the comments in the header of the script for more details on usage and more docs.

References #9.