chore(): update typescript linting
Previously the TypeScript linter (tslint) we used was using the default rules or the rules defined in `tslint.json` in the root of the project. Gulp tasks now explicitly reference the `tslint.json` in the root of the repo. A local workspace settings file (.vscode/settings.json) also configures the tslint [vscode extension](https://github.com/Microsoft/vscode-tslint) to use the project's `tslint.json` file.