chore(vet): fix vet to fail on errors
The addition of the gulp task 'live-dev' introduced a regression in our build process. The task 'live-dev' has to swallow all errors so the process does not terminate, rather it keeps running to watch for file changes for continuous development.

However when vetting code, this means the `emitError` on the TSLINT utility must be set to false telling it to not emit any errors. While this works for `live-dev`, the vetting gulp tasks now write out the linting errors found, but the process does not exit as an error. The effects of this is our CD process does not detect these errors are build errors which it should.

This fix adds an optional `--noExit` argument to the vetting tasks. If set, the TSLINT utility will write the error, but not emit an error making it work for the live-dev. The default state though is to fail the process, resolving the issue with our CD process.

Closes #192. Closes #211.