chore(): switch from tsd to typings
The TSD project for acquiring TypeScrupt type definitions is now legacy and reached end of life. The Typings project has taken it's place. This commit replaces the usage of TSD with Typings. Unfortunately Typings does not have a way to specify where the type definitions are kept in the project, so we had to move them from the `/build/typings` to the root of the project in `/typings`.

Closes #215. Closes #238.