fix(pivot): fixed Pivot demo to display contextual menu in ellipsis
Pivot demo did not display properly the contextual menu that is part of the ellipsis component. This has been identified as a bug within Fabric UI ([Pivot: ContextualMenu not showing](https://github.com/OfficeDev/Office-UI-Fabric/issues/508)).
Demo now has a workaround that issue until it will be solved in Fabric UI. The workaround is overriding the `overflow-x` property of the Pivot element.