fix(breadcrumb): missing responsiveness
Breaking changes:
- component does not allow transclusion of the `uif-breadcrumb-link` elements.
- breadcrumb items has to be passed as expression evaluating to an array of elements containing link text and href.

Added missing responsiveness to breadcrumb. Based on browser width, 4 or 2 last breadcrumb items are displayed.

Rest of the elements is placed in overflow menu, which is accessible via the ellipsis icon.

In order to add this feature, internal implementation has been changed as well as input of the elements to display.

Closes #342.