feat(panel): add uifIslightDismiss attribute
Add `uif-is-light-dismiss` as an attribute of the Panel directive to allow the setting to determine whether clicking on the background overlay closes the panel. Default value is false. When not set, or set to false, clicking on the background overlay has no effect. When set to true, clicking on the background overlay closes the panel.

Closes #434.